+++ src/components/Main/index.js
@@ -13,15 +13,23 @@ import CreateCutting from '../Cuttings/CreateCuttingContainer';
 function Main() {
   return (
     <div>
-      <Route path="/about/" exact component={About} />
-      <Route path="/events/" exact component={Events} />
-      <Route path="/events/:id" exact component={EventDetails} />
-      <Route path="/create-event/" exact component={CreateEventContainer} />
-      <Route path="/signup/" exact component={SignupContainer} />
-      <Route path="/login/" exact component={LoginContainer} />
-      <Route path="/cuttings/" exact component={Cuttings} />
-      <Route path="/cuttings/:id" exact component={CuttingDetails} />
-      <Route path="/create-cutting/" exact component={CreateCutting} />
+      <div>
+        <Route path="/events/" exact component={Events} />
+        <Route path="/events/:id" exact component={EventDetails} />
+        <Route path="/create-event/" exact component={CreateEventContainer} />
+      </div>
+      <div>
+        <Route path="/cuttings/" exact component={Cuttings} />
+        <Route path="/cuttings/:id" exact component={CuttingDetails} />
+        <Route path="/create-cutting/" exact component={CreateCutting} />
+      </div>
+      <div>
+        <Route path="/about/" exact component={About} />
+      </div>
+      <div>
+        <Route path="/signup/" exact component={SignupContainer} />
+        <Route path="/login/" exact component={LoginContainer} />
+      </div>
     </div>
   )
 }
