+++ src/components/Navigation/index.js
@@ -6,35 +6,23 @@ import { withTranslation } from 'react-i18next';
 class Navigation extends Component {
   render() {
 
-    const { t, i18n } = this.props
+    const { t } = this.props
 
-    const changeLanguage = lng => {
-      i18n.changeLanguage(lng);
-    };
 
     return (
       <div>
-        <div>
-          <nav>
-            {/* <SelectLanguage /> */}
-
-            <nav>
-              <button onClick={() => changeLanguage('en')} className="btn btn--yellow">EN</button>
-            </nav>
-            <nav>
-              <button onClick={() => changeLanguage('nl')} className="btn btn--yellow">NL</button>
-            </nav>
-
-          </nav>
-
+        <div className="navigation__main">
           <nav>
             <Link to="/events/"><button className="btn btn--yellow">{t('navigation.events')}</button></Link>
             <Link to="/cuttings/"><button className="btn btn--yellow">{t('navigation.cuttings')}</button></Link>
             <Link to="/about/"><button className="btn btn--yellow">{t('navigation.about')}</button></Link>
+          </nav>
+
+          <div className="navigation__user">
             <Link to="/signup/"><button className="btn btn--yellow">{t('navigation.signup')}</button></Link>
             <Link to="/login/"><button className="btn btn--yellow">{t('navigation.login')}</button></Link>
+          </div>
 
-          </nav>
         </div>
 
       </div>
