+++ src/css/style.css
@@ -20,12 +20,15 @@ html {
   text-decoration: none; }
 
 body {
-  padding: 3rem; }
+  padding: 1rem; }
 
 .container {
   display: grid;
-  grid-template-rows: 80vh min-content 40vw repeat(3, min-content);
-  grid-template-columns: [sidebar-start] 8rem [sidebar-end full-start] minmax(6rem, 1fr) [center-start] repeat(8, [col-start] minmax(min-content, 14rem) [col-end]) [center-end] minmax(6rem, 1fr) [Full-end]; }
+  grid-template-rows: 25vh repeat(2, min-content);
+  grid-template-columns: [navigation-start] 20rem [navigation-end full-start] 1fr [center-start] repeat(8, [col-start] minmax(min-content, 14rem) [col-end]) [center-end] 1fr [full-end]; }
+  .container > * {
+    padding: 40px;
+    font-size: 3rem; }
 
 body {
   font-weight: 300;
@@ -115,91 +118,33 @@ body {
   box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.15);
   transform: translateY(0); }
 
-.toggle-switch {
-  position: relative;
-  width: 75px;
-  display: inline-block;
-  vertical-align: middle;
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  text-align: left; }
-  .toggle-switch-checkbox {
-    display: none; }
-  .toggle-switch-label {
-    display: block;
-    overflow: hidden;
-    cursor: pointer;
-    border: 0 solid #ccc;
-    border-radius: 20px;
-    margin: 0; }
-  .toggle-switch-inner {
-    display: block;
-    width: 200%;
-    margin-left: -100%;
-    transition: margin 0.3s ease-in 0s; }
-    .toggle-switch-inner:before, .toggle-switch-inner:after {
-      display: block;
-      float: left;
-      width: 50%;
-      height: 34px;
-      padding: 0;
-      line-height: 34px;
-      font-size: 14px;
-      color: white;
-      font-weight: bold;
-      box-sizing: border-box; }
-    .toggle-switch-inner:before {
-      content: attr(data-yes);
-      text-transform: uppercase;
-      padding-left: 10px;
-      background-color: #f90;
-      color: #fff; }
-  .toggle-switch-disabled {
-    background-color: #ccc;
-    cursor: not-allowed; }
-    .toggle-switch-disabled:before {
-      background-color: #ccc;
-      cursor: not-allowed; }
-  .toggle-switch-inner:after {
-    content: attr(data-no);
-    text-transform: uppercase;
-    padding-right: 10px;
-    background-color: #ccc;
-    color: #fff;
-    text-align: right; }
-  .toggle-switch-switch {
-    display: block;
-    width: 24px;
-    margin: 5px;
-    background: #fff;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    right: 40px;
-    border: 0 solid #ccc;
-    border-radius: 20px;
-    transition: all 0.3s ease-in 0s; }
-  .toggle-switch-checkbox:checked + .toggle-switch-label .toggle-switch-inner {
-    margin-left: 0; }
-  .toggle-switch-checkbox:checked + .toggle-switch-label .toggle-switch-switch {
-    right: 0px; }
-  .toggle-switch.small-switch {
-    width: 40px; }
-    .toggle-switch.small-switch .toggle-switch-inner:after, .toggle-switch.small-switch .toggle-switch-inner:before {
-      content: "";
-      height: 20px;
-      line-height: 20px; }
-    .toggle-switch.small-switch .toggle-switch-switch {
-      width: 16px;
-      right: 20px;
-      margin: 2px; }
-  @media screen and (max-width: 991px) {
-    .toggle-switch {
-      transform: scale(0.9); } }
-  @media screen and (max-width: 767px) {
-    .toggle-switch {
-      transform: scale(0.825); } }
-  @media screen and (max-width: 575px) {
-    .toggle-switch {
-      transform: scale(0.75); } }
+.about {
+  background-color: yellow;
+  grid-column: full-start / full-end; }
+
+.body {
+  background-color: yellow; }
+
+.cuttings {
+  grid-column: center-start / center-end;
+  background-color: indianred; }
+
+.events {
+  grid-column: center-start / center-end;
+  background-color: greenyellow; }
+
+.footer {
+  background-color: hotpink;
+  grid-column: full-start / full-end; }
+
+.header {
+  background-color: blue;
+  grid-column: full-start / full-end; }
+
+.navigation {
+  background-color: red;
+  grid-column: navigation-start / navigation-end;
+  grid-row: 1 / -1; }
+
+.users {
+  grid-column: center-start / center-end; }
