+++ src/sass/base/_base.scss
@@ -24,12 +24,18 @@ html {
 
 body {
   
-   padding: 3rem;
+   padding: 1rem;
 }
 
 .container {
   display: grid;
-  grid-template-rows: 80vh min-content 40vw repeat(3, min-content);
-  grid-template-columns: [sidebar-start] 8rem [sidebar-end full-start] minmax(6rem, 1fr)
-  [center-start] repeat(8, [col-start] minmax(min-content, 14rem) [col-end]) [center-end] minmax(6rem, 1fr) [Full-end];
+  grid-template-rows: 25vh repeat(2, min-content);
+  grid-template-columns: [navigation-start] 20rem [navigation-end full-start] 1fr [center-start] repeat(8,
+  [col-start] minmax(min-content, 14rem) [col-end])[center-end] 1fr [full-end];
+
+  & > * {
+    padding: 40px;
+    font-size: 3rem;
+  }
+
 }
