+++ src/sass/layout/_about.scss
@@ -1,0 +1,4 @@
+.about {
+  background-color: yellow;
+  grid-column: full-start / full-end;
+}
