+++ src/sass/layout/_body.scss
@@ -1,0 +1,5 @@
+.body {
+  background-color: yellow;
+  // grid-column: full-start / full-end;
+
+}
