+++ src/sass/layout/_cuttings.scss
@@ -1,0 +1,4 @@
+.cuttings {
+  grid-column: center-start / center-end;
+  background-color: indianred;
+}
