+++ src/sass/layout/_events.scss
@@ -1,0 +1,4 @@
+.events {
+  grid-column: center-start / center-end;
+  background-color: greenyellow;
+}
