+++ src/sass/layout/_footer.scss
@@ -1,0 +1,4 @@
+.footer {
+  background-color: hotpink;
+    grid-column: full-start / full-end;
+}
