+++ src/sass/layout/_header.scss
@@ -1,0 +1,4 @@
+.header {
+  background-color: blue;
+  grid-column: full-start / full-end;
+}
