+++ src/sass/layout/_navigation.scss
@@ -1,0 +1,5 @@
+.navigation {
+  background-color: red;
+  grid-column: navigation-start / navigation-end;
+  grid-row: 1 / -1;
+}
