+++ src/sass/layout/_users.scss
@@ -1,0 +1,3 @@
+.users {
+  grid-column: center-start / center-end;
+}
