+++ src/sass/main.scss
@@ -6,3 +6,11 @@
 @import "components/about";
 @import "components/buttons";
 
+@import "layout/about";
+@import "layout/body";
+@import "layout/cuttings";
+@import "layout/events";
+@import "layout/footer";
+@import "layout/header";
+@import "layout/navigation";
+@import "layout/users";
